app/requirements.txt
fastapi==0.95.1
#redis==4.3.3
uvicorn==0.18.2
redis-py-cluster
